Lens types package
