+++ packages/kbn-lens-types/README.md
@@ -1,0 +1,3 @@
+# @kbn/lens-types
+
+A place to put Lens-related types for sharing outside Lens. This helps us avoid circular type dependencies.
