+++ packages/kbn-lens-types/index.ts
@@ -1,0 +1,25 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0 and the Server Side Public License, v 1; you may not use this file except
+ * in compliance with, at your election, the Elastic License 2.0 or the Server
+ * Side Public License, v 1.
+ */
+
+export type {
+  DataType,
+  SortingHint,
+  TimeScaleUnit,
+  OperationMetadata,
+  Operation,
+  BaseIndexPatternColumn,
+  FieldBasedIndexPatternColumn,
+  ValueFormatConfig,
+  FormattedIndexPatternColumn,
+  ReferenceBasedIndexPatternColumn,
+  GenericIndexPatternColumn,
+  FormBasedLayer,
+  FormBasedPersistedState,
+  OperationType,
+  IncompleteColumn,
+} from './form_based';
