+++ packages/kbn-lens-types/kibana.jsonc
@@ -1,0 +1,5 @@
+{
+  "type": "shared-common",
+  "id": "@kbn/lens-types",
+  "owner": "@elastic/kibana-visualizations"
+}
