+++ x-pack/plugins/lens/public/embeddable/embeddable_component.tsx
@@ -20,9 +20,9 @@ import {
   IEmbeddable,
   useEmbeddableFactory,
 } from '@kbn/embeddable-plugin/public';
+import type { FormBasedPersistedState } from '@kbn/lens-types';
 import type { LensByReferenceInput, LensByValueInput } from './embeddable';
 import type { Document } from '../persistence';
-import type { FormBasedPersistedState } from '../datasources/form_based/types';
 import type { TextBasedPersistedState } from '../datasources/text_based/types';
 import type { XYState } from '../visualizations/xy/types';
 import type {
